Rename main.py to model.py
